Rev Defs and Import Organization.
I updated the cache data to 167.
All the Revenant definitions are added cache and server side.

I organized the imports so that removed 100+ warnings.

There was also some other resource leaks that were fixed, which in some cases can cause memory leaks.

Added a few random things to help with my process with adding revs.

They just need spawned and mechanics.

+++ client and cache/client/src/com/osps/client/Sprite.java
@@ -1,5 +1,8 @@
 package com.osps.client;
-import java.awt.*;
+import java.awt.Component;
+import java.awt.Image;
+import java.awt.MediaTracker;
+import java.awt.Toolkit;
 import java.awt.image.PixelGrabber;
 
 import javax.swing.ImageIcon;
@@ -638,8 +641,8 @@ public final class Sprite extends DrawingArea {
 		try {
 			int j2 = -l1 / 2;
 			int k2 = -i / 2;
-			int l2 = (int) (Math.sin((double) j / 326.11000000000001D) * 65536D);
-			int i3 = (int) (Math.cos((double) j / 326.11000000000001D) * 65536D);
+			int l2 = (int) (Math.sin(j / 326.11000000000001D) * 65536D);
+			int i3 = (int) (Math.cos(j / 326.11000000000001D) * 65536D);
 			l2 = l2 * k >> 8;
 			i3 = i3 * k >> 8;
 			int j3 = (i2 << 16) + (k2 * l2 + j2 * i3);
