+++ client and cache/client/src/com/osps/client/Stream.java
@@ -229,8 +229,8 @@ public final class Stream extends NodeSub {
 	}
 
 	public long readQWord() {
-		long l = (long) readDWord() & 0xffffffffL;
-		long l1 = (long) readDWord() & 0xffffffffL;
+		long l = readDWord() & 0xffffffffL;
+		long l1 = readDWord() & 0xffffffffL;
 		return (l << 32) + l1;
 	}
