+++ client and cache/client/src/com/osps/client/draw/Rasterizer.java
@@ -1,7 +1,5 @@
 package com.osps.client.draw;
 
-import java.util.concurrent.Executors;
-
 import com.osps.client.Client;
 import com.osps.client.Configuration;
 import com.osps.client.cache.StreamLoader;
@@ -1824,11 +1822,11 @@ public final class Rasterizer extends DrawingArea {
 		int j = 0;
 		for(int k = 0; k < 512; k++)
 		{
-			double d1 = (double)(k / 8) / 64D + 0.0078125D;
-			double d2 = (double)(k & 7) / 8D + 0.0625D;
+			double d1 = k / 8 / 64D + 0.0078125D;
+			double d2 = (k & 7) / 8D + 0.0625D;
 			for(int k1 = 0; k1 < 128; k1++)
 			{
-				double d3 = (double)k1 / 128D;
+				double d3 = k1 / 128D;
 				double d4 = d3;
 				double d5 = d3;
 				double d6 = d3;
@@ -1911,9 +1909,9 @@ public final class Rasterizer extends DrawingArea {
 
 	private static int method373(int i, double d)
 	{
-		double d1 = (double)(i >> 16) / 256D;
-		double d2 = (double)(i >> 8 & 0xff) / 256D;
-		double d3 = (double)(i & 0xff) / 256D;
+		double d1 = (i >> 16) / 256D;
+		double d2 = (i >> 8 & 0xff) / 256D;
+		double d3 = (i & 0xff) / 256D;
 		d1 = Math.pow(d1, d);
 		d2 = Math.pow(d2, d);
 		d3 = Math.pow(d3, d);
@@ -6223,8 +6221,8 @@ public static void drawMaterializedTriangle(int y1, int y2, int y3, int x1, int 
 
 		for(int k = 0; k < 2048; k++)
 		{
-			anIntArray1470[k] = (int)(65536D * Math.sin((double)k * 0.0030679614999999999D));
-			anIntArray1471[k] = (int)(65536D * Math.cos((double)k * 0.0030679614999999999D));
+			anIntArray1470[k] = (int)(65536D * Math.sin(k * 0.0030679614999999999D));
+			anIntArray1471[k] = (int)(65536D * Math.cos(k * 0.0030679614999999999D));
 		}
 
 	}
