+++ client and cache/client/src/com/osps/client/map/ObjectManager.java
@@ -517,7 +517,7 @@ public final class ObjectManager {
 	private static int method172(int i, int j)
 	{
 		int k = (method176(i + 45365, j + 0x16713, 4) - 128) + (method176(i + 10294, j + 37821, 2) - 128 >> 1) + (method176(i, j, 1) - 128 >> 2);
-		k = (int)((double)k * 0.29999999999999999D) + 35;
+		k = (int)(k * 0.29999999999999999D) + 35;
 		if(k < 10)
 			k = 10;
 		else
